mod 보안
- 삭제 요청시 주문자와 일치하는 경우에만 삭제 가능하도록 수정

+++ src/main/java/com/galid/commerce/domains/order/service/MyOrderService.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -61,7 +62,13 @@ public class MyOrderService {
         return myOrderDetailsDto;
     }
 
-    public void deleteMyOrder(Long orderId) {
-        orderRepository.deleteById(orderId);
+    public void deleteMyOrder(Long ordererId, Long orderId) {
+        OrderEntity order  = orderRepository.findById(orderId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 주문입니다."));
+
+        if (ordererId == order.getOrderer().getMemberId())
+            orderRepository.deleteById(orderId);
+        else
+            throw new IllegalStateException("주문자와 삭제 요청자가 일치하지 않습니다.");
     }
 }
